IssueNo:https://gitee.com/openharmony/graphic_ui/issues/I3ZDVW Description:support cmath sin function Sig:graphic Feature or Bugfix:Bugfix Binary Source:No Signed-off-by: wangtiantian <wangtiantian19@huawei.com>

+++ interfaces/kits/gfx_utils/graphic_math.h
@@ -77,7 +77,7 @@ static constexpr float ATAN2_P7 = -0.0443360;
 static constexpr float RADIAN_TO_ANGLE = 57.295779513;
 static constexpr float UI_FLT_EPSILON = 1E-5;
 
-float Sin(int16_t angle);
+float Sin(float angle);
 uint16_t FastAtan2(int16_t x, int16_t y);
 float Sqrt(float x);
